firwmare_v1: implement console protocol

+++ firmware_v1/canmgr.h
@@ -1,5 +1,5 @@
 
-void canmgr_setup (uint32_t addr);
+void canmgr_setup (uint8_t addr);
 void canmgr_finalize (void);
 void canmgr_update (void);
